Move code in spyder/plugin/__init__.py to spyder/api/plugins.py
